# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RAG pipeline over Banner release notes and Standard Operating Procedures (SOPs), backed by Azure OpenAI and Azure Cognitive Search.",
        "title": "go-omnivore-rag API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/chat/ask": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Route a student message to Banner or SOP backend based on intent/source",
                "parameters": [
                    {
                        "description": "Chat message payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.chatAskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/adapter.AdapterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/chat/intent": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Classify the intent of a student message (6 intents)",
                "parameters": [
                    {
                        "description": "Message to classify",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.messageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.intentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/chat/sentiment": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Analyze the sentiment of a student message",
                "parameters": [
                    {
                        "description": "Message to analyze",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.messageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.sentimentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "adapter.AdapterResponse": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "confidence": {
                    "description": "Confidence score from sources[0].score (0–1); 0.0 if no sources returned",
                    "type": "number"
                },
                "escalate": {
                    "description": "true when confidence < 0.5 or retrieval_count == 0",
                    "type": "boolean"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/adapter.AdapterSource"
                    }
                }
            }
        },
        "adapter.AdapterSource": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "sop_number": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "api.chatAskRequest": {
            "type": "object",
            "required": [
                "message",
                "session_id"
            ],
            "properties": {
                "intent": {
                    "description": "Optional intent hint: RegistrationBanner | FinanceBanner | TranscriptSop | HoldsSop | ReleaseSummary | General",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "source": {
                    "description": "Optional routing override: general | student | finance | sop | auto",
                    "type": "string",
                    "enum": [
                        "general",
                        "student",
                        "finance",
                        "sop",
                        "auto"
                    ]
                }
            }
        },
        "api.intentResponse": {
            "type": "object",
            "properties": {
                "confidence": {
                    "description": "Normalized score 0–1",
                    "type": "number"
                },
                "intent": {
                    "description": "One of RegistrationBanner | FinanceBanner | TranscriptSop | HoldsSop | ReleaseSummary | General",
                    "type": "string"
                }
            }
        },
        "api.messageRequest": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "api.sentimentResponse": {
            "type": "object",
            "properties": {
                "score": {
                    "description": "Normalized score 0–1",
                    "type": "number"
                },
                "sentiment": {
                    "description": "One of Positive | Neutral | Frustrated",
                    "type": "string"
                }
            }
        },
...
    }
}